removed everything related to quarkus-deployment-dependency.graph
